Introduced BaseFhirQueryGenerator
